requirements.txt
asgiref==3.8.1
Django==5.2
django-filter==25.1
django-rest-knox==5.0.2
djangorestframework==3.16.0
psycopg2-binary==2.9.10
Pygments==2.19.1
python-dotenv==1.1.1
sqlparse==0.5.3
